pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "google-genai>=1.28.0",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sift-stack-py>=0.8.0",
]
